requirements.txt
requests~=2.28.2
discord~=2.2.2
openai~=0.27.2
python-dotenv~=1.00
pyaudio~=0.2.13
SpeechRecognition~=3.10.0